YummyRecipesApi/requirements.txt
aniso8601==3.0.2
atomicwrites==1.1.5
attrs==18.1.0
Django==2.1
django-environ==0.4.5
django-graphql-jwt==0.1.10
graphene==2.1.3
graphene-django==2.1.0
graphql-core==2.1
graphql-relay==0.4.5
iso8601==0.1.12
more-itertools==4.3.0
pluggy==0.7.1
promise==2.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.5.4
PyJWT==1.6.4
pytest==3.7.1
pytz==2018.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.4
